sloane-frontend-package/ai-phone-service-v2/ai-phone-service-v2/requirements.txt
fastapi==0.115.0
uvicorn==0.24.0
twilio==8.0.0
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.1
tensorflow==2.14.0
google-api-python-client==2.100.0
google-auth-oauthlib==1.0.0
python-dotenv==1.0.0
pydantic==2.4.2
